new version of upgrade.sh

+++ upgrade.sh
@@ -14,13 +14,24 @@ if [[ $response =~ ^(yes|y|Y)$ ]]; then
     [[ -d $wwwdir ]] || sudo mkdir -p $wwwdir
     [[ -d $installdir ]] || mkdir -p $installdir
     [[ -d /home/pi/Videos ]] || mkdir -p /home/pi/Videos
+    [[ ! -d $unpackdir ]] || rm -r $unpackdir
+    [[ ! -f $downloaddir/RPiMS.zip ]] || rm $downloaddir/RPiMS.zip
+    curl -sS $repourl -L -o $downloaddir/RPiMS.zip
+    unzip  $downloaddir/RPiMS.zip -d $downloaddir
 
-    #curl -sS $repourl -L -o $downloaddir/RPiMS.zip
-    #unzip  $downloaddir/RPiMS.zip -d $downloaddir
     #sudo cp -R $unpackdir/www/* $wwwdir
     #sudo cp $unpackdir/RPiMS/* $installdir
     #sudo chmod u+x $installdir/*.py $installdir/*.sh
-    #rm $downloaddir/RPiMS.zip
+
+    #[[ ! -d $unpackdir ]] || rm -r $unpackdir
+    #[[ ! -f $downloaddir/RPiMS.zip ]] || rm $downloaddir/RPiMS.zip
+
+    echo ""
+    echo "-------------------------------------"
+    echo "Upgrade successfully completed !"
+    echo "-------------------------------------"
+    echo ""
+
 fi
 
 echo "Do you want to upgrade the Operating System ?"
@@ -30,6 +41,18 @@ if [[ $response =~ ^(yes|y|Y)$ ]]; then
     sudo apt-get -y update
     sudo apt-get -y upgrade
     sudo apt-get -y autoremove
+
+    echo ""
+    echo "-------------------------------------"
+    echo "Upgrade successfully completed !"
+    echo "-------------------------------------"
+    echo ""
+fi
+
+echo "Do you want to upgrade the Python library ?"
+read -r -p "$1 [y/N] " response < /dev/tty
+if [[ $response =~ ^(yes|y|Y)$ ]]; then
+    echo "Greats ! The upgrade OS has started."
 
     sudo python3 -m pip install --upgrade pip setuptools wheel
     sudo -H pip3 install --upgrade RPi.bme280 smbus2 redis hiredis pid PyYAML luma.oled luma.lcd adafruit-circuitpython-ads1x15 rshell pyusb
